[10126]: Move the help command stuff into the new cmd area.
